comp_bot/requirements.txt
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
protobuf==6.33.3
pyfrc==2026.0.1
PyQt6==6.10.2
robotpy==2026.1.1
robotpy-apriltag==2026.1.1
robotpy-cli==2026.0.1
robotpy-commands-v2==2026.1.1
robotpy-cscore==2026.1.1
robotpy-hal==2026.1.1
robotpy-halsim-ds-socket==2026.1.1
robotpy-halsim-gui==2026.1.1
robotpy-halsim-ws==2026.1.1
robotpy-installer==2026.0.1
robotpy-native-apriltag==2026.1.1
robotpy-native-ntcore==2026.1.1
robotpy-native-romi==2026.1.1
robotpy-native-wpihal==2026.1.1
robotpy-native-wpilib==2026.1.1
robotpy-native-wpimath==2026.1.1
robotpy-native-wpinet==2026.1.1
robotpy-native-wpiutil==2026.1.1
robotpy-native-xrp==2026.1.1
robotpy-navx==2026.0.0b1
robotpy-pathplannerlib==2025.2.7
robotpy-rev==2026.0.0
robotpy-romi==2026.1.1
robotpy-urcl==2026.0.0
robotpy-wpilib-utilities==2026.0.1
robotpy-wpimath==2026.1.1
robotpy-wpinet==2026.1.1
robotpy-wpiutil==2026.1.1
wpilib==2026.1.1
